Move frontend/packages/dev-console/src/components/projects -> frontend/packages/console-shared/src/components/projects
