+++ frontend/packages/console-shared/src/components/projects/NamespacedPage.scss
