+++ frontend/packages/console-shared/src/components/projects/details/__tests__/ProjectDetailsPage.spec.tsx
