+++ frontend/packages/helm-plugin/src/components/list-page/HelmReleaseListPage.tsx
@@ -3,12 +3,12 @@ import { Button } from '@patternfly/react-core';
 import Helmet from 'react-helmet';
 import { useTranslation, Trans } from 'react-i18next';
 import { RouteComponentProps } from 'react-router';
-import NamespacedPage, {
-  NamespacedPageVariants,
-} from '@console/dev-console/src/components/NamespacedPage';
-import CreateProjectListPage from '@console/dev-console/src/components/projects/CreateProjectListPage';
 import { withStartGuide } from '@console/internal/components/start-guide';
 import { PageHeading } from '@console/internal/components/utils';
+import CreateProjectListPage from '@console/shared/src/components/projects/CreateProjectListPage';
+import NamespacedPage, {
+  NamespacedPageVariants,
+} from '@console/shared/src/components/projects/NamespacedPage';
 import HelmReleaseList from './HelmReleaseList';
 
 type HelmReleaseListPageProps = RouteComponentProps<{ ns: string }>;
