+++ frontend/packages/pipelines-plugin/src/components/pipelines/__tests__/PipelineTabbedPage.spec.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import NamespacedPage from '@console/dev-console/src/components/NamespacedPage';
-import CreateProjectListPage from '@console/dev-console/src/components/projects/CreateProjectListPage';
 import { MultiTabListPage, useFlag } from '@console/shared';
+import CreateProjectListPage from '@console/shared/src/components/projects/CreateProjectListPage';
+import NamespacedPage from '@console/shared/src/components/projects/NamespacedPage';
 import { PipelinesPage } from '../PipelinesPage';
 import PipelineTabbedPage, { PageContents } from '../PipelineTabbedPage';
