+++ frontend/packages/pipelines-plugin/src/components/pipelines/__tests__/PipelinesPage.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import CreateProjectListPage from '@console/dev-console/src/components/projects/CreateProjectListPage';
 import { referenceForModel } from '@console/internal/module/k8s';
+import CreateProjectListPage from '@console/shared/src/components/projects/CreateProjectListPage';
 import { PipelineModel } from '../../../models';
 import { PipelinesPage } from '../PipelinesPage';
 import PipelinesResourceList from '../PipelinesResourceList';
